Copilot: refactoring phase 6
