--- cmd/oceano-state-manager/library.go
@@ -1,8 +1,0 @@
-package main
-
-import internallibrary "github.com/alemser/oceano-player/internal/library"
-
-type Library = internallibrary.Library
-type CollectionEntry = internallibrary.CollectionEntry
-
-var Open = internallibrary.Open
